docstrings comments added

+++ vocabulary.py
@@ -1,3 +1,6 @@
+"""
+File containing the functions to interact with the CosmosDB database where the words are stored and retrieved from
+"""
 import os
 from dotenv import load_dotenv
 from azure.cosmos import exceptions, CosmosClient, PartitionKey
@@ -5,7 +8,10 @@ from azure.identity import DefaultAzureCredential
 import uuid
 from languageDetection import detect_language_code
 
-#Creating connection to CosmosDB
+
+"""
+Creating connection to CosmosDB
+"""
 load_dotenv()
 credential = DefaultAzureCredential()
 cosmos_endpoint = os.getenv("COSMOS_ACCOUNT_URI")
@@ -13,24 +19,40 @@ cosmos_client = CosmosClient(cosmos_endpoint, credential=credential)
 dictionary_database = cosmos_client.get_database_client("dictionary")
 words_container = dictionary_database.get_container_client("words")
 
-#Function to import the list of words in a local array from cosmosDB given the user ID
+
 def get_all_words(user_id):
+    """
+    Function to get all the words from the dictionary for a given user
+
+    args:
+    user_id: the user id to get the words for
+
+    returns:
+    words: a list of all the words in the dictionary for the user
+    """
     items = list(words_container.query_items(query="SELECT * FROM c WHERE c.user_id = @user_id", parameters=[dict(name="@user_id", value=user_id)]))
     words = [[item['text'], item['translation']['text']] for item in items]
     #words = [item['translation']['text'] for item in items]
     return words
 
 
-#Function to save a new word to the CosmosDB - id generated locally with uuid and user_id hardcoded for now 
-#If an element with the same id already exists for all the partition keys then I create a new id with uuid
-#If there is a duplicate, the word will not be saved
 def save_word(text, translation):
+    """
+    Function to save a new word to the dictionary in CosmosDB
+
+    args:
+    text: the word in the native language
+    translation: the translation of the word
+
+    returns:
+    message: a message indicating if the word has been added or if it already exists
+    """
     text = text.lower()
     translation = translation.lower()
     unique_id = str(uuid.uuid4())
     language_code = detect_language_code(text)
     translation_language_code = detect_language_code(translation)
-
+    
     words_with_same_id = list(words_container.query_items(query="SELECT * FROM c WHERE c.id = @id", parameters=[dict(name="@id", value=unique_id)], enable_cross_partition_query=True))
     num_words_with_same_id = len(words_with_same_id)
     while num_words_with_same_id != 0:
@@ -53,17 +75,28 @@ def save_word(text, translation):
         return "The word has been added to the dictionary"
 
 
-#Function to delete a given word (native language and its translation) from cosmosDB
 def delete_word(text, translation):
+    """
+    Function to delete a word from the dictionary in CosmosDB given the word and its translation
+
+    args:
+    text: the word in the native language
+    translation: the translation of the word
+
+    returns:
+    binary: a boolean indicating if the word has been deleted or not
+    """
     text = text.lower()
     translation = translation.lower()
     try:
         items = list(words_container.query_items(query="SELECT * FROM c WHERE c.user_id = '553fcaa0-2530-472c-9126-ffec24c62a6c' AND c.text = @text AND c.translation.text = @translation", parameters=[dict(name="@text", value=text), dict(name="@translation", value=translation)]))
-        for item in items:
-            words_container.delete_item(item, partition_key=item['user_id']) 
-        return "The word has been deleted from the dictionary"
+        if len(items) != 0:
+            print(items)
+            for item in items:
+                words_container.delete_item(item, partition_key=item['user_id']) 
+            return True
+        else:
+            return False
     except Exception as e:
-        print("Word not found in the CosmosDB dictionary " + str(e))
-        return "The word is not in the dictionary"
-
-
+        print("An error occurred " + str(e))
+        return False
